delete duplication codes
