--- shardingsphere-agent/shardingsphere-agent-plugins/shardingsphere-agent-plugin-tracing/shardingsphere-agent-tracing-jaeger/src/main/java/org/apache/shardingsphere/agent/plugin/tracing/jaeger/advice/CommandExecutorTaskAdvice.java
@@ -1,67 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.agent.plugin.tracing.jaeger.advice;
-
-import io.opentracing.Scope;
-import io.opentracing.tag.Tags;
-import io.opentracing.util.GlobalTracer;
-import lombok.SneakyThrows;
-import org.apache.shardingsphere.agent.api.advice.AdviceTargetObject;
-import org.apache.shardingsphere.agent.api.advice.InstanceMethodAroundAdvice;
-import org.apache.shardingsphere.agent.api.result.MethodInvocationResult;
-import org.apache.shardingsphere.agent.plugin.tracing.jaeger.constant.JaegerConstants;
-import org.apache.shardingsphere.agent.plugin.tracing.jaeger.span.JaegerErrorSpan;
-import org.apache.shardingsphere.infra.executor.kernel.model.ExecutorDataMap;
-import org.apache.shardingsphere.proxy.backend.communication.jdbc.connection.BackendConnection;
-import org.apache.shardingsphere.proxy.frontend.command.CommandExecutorTask;
-
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
-
-/**
- * Command executor task advice.
- */
-public final class CommandExecutorTaskAdvice implements InstanceMethodAroundAdvice {
-    
-    private static final String OPERATION_NAME = "/ShardingSphere/rootInvoke/";
-    
-    @Override
-    public void beforeMethod(final AdviceTargetObject target, final Method method, final Object[] args, final MethodInvocationResult result) {
-        Scope scope = GlobalTracer.get().buildSpan(OPERATION_NAME)
-                .withTag(Tags.COMPONENT.getKey(), JaegerConstants.COMPONENT_NAME)
-                .startActive(true);
-        ExecutorDataMap.getValue().put(JaegerConstants.ROOT_SPAN, scope.span());
-    }
-    
-    @Override
-    @SneakyThrows
-    public void afterMethod(final AdviceTargetObject target, final Method method, final Object[] args, final MethodInvocationResult result) {
-        ExecutorDataMap.getValue().remove(JaegerConstants.ROOT_SPAN);
-        Field field = CommandExecutorTask.class.getDeclaredField("backendConnection");
-        field.setAccessible(true);
-        BackendConnection connection = (BackendConnection) field.get(target);
-        Scope scope = GlobalTracer.get().scopeManager().active();
-        scope.span().setTag(JaegerConstants.ShardingSphereTags.CONNECTION_COUNT.getKey(), connection.getConnectionSize());
-        scope.close();
-    }
-    
-    @Override
-    public void onThrowing(final AdviceTargetObject target, final Method method, final Object[] args, final Throwable throwable) {
-        JaegerErrorSpan.setError(GlobalTracer.get().activeSpan(), throwable);
-    }
-}
