+++ shardingsphere-agent/shardingsphere-agent-plugins/shardingsphere-agent-plugin-tracing/shardingsphere-agent-tracing-jaeger/src/main/java/org/apache/shardingsphere/agent/plugin/tracing/jaeger/definition/JaegerPluginDefinitionService.java
@@ -17,54 +17,22 @@
 
 package org.apache.shardingsphere.agent.plugin.tracing.jaeger.definition;
 
-import java.util.Collection;
-import net.bytebuddy.matcher.ElementMatchers;
 import org.apache.shardingsphere.agent.api.point.PluginInterceptorPoint;
+import org.apache.shardingsphere.agent.plugin.tracing.opentracing.definition.OpenTracingPluginDefinitionService;
 import org.apache.shardingsphere.agent.spi.definition.AbstractPluginDefinitionService;
+
+import java.util.Collection;
 
 /**
  * Jaeger plugin definition service.
  */
 public final class JaegerPluginDefinitionService extends AbstractPluginDefinitionService {
     
-    private static final String COMMAND_EXECUTOR_TASK_ENHANCE_CLASS = "org.apache.shardingsphere.proxy.frontend.command.CommandExecutorTask";
-    
-    private static final String COMMAND_EXECUTOR_METHOD_NAME = "executeCommand";
-    
-    private static final String COMMAND_EXECUTOR_TASK_ADVICE_CLASS = "org.apache.shardingsphere.agent.plugin.tracing.jaeger.advice.CommandExecutorTaskAdvice";
-    
-    private static final String SQL_PARSER_ENGINE_ENHANCE_CLASS = "org.apache.shardingsphere.infra.parser.ShardingSphereSQLParserEngine";
-    
-    private static final String SQL_PARSER_ENGINE_METHOD_NAME = "parse";
-    
-    private static final String SQL_PARSER_ENGINE_ADVICE_CLASS = "org.apache.shardingsphere.agent.plugin.tracing.jaeger.advice.SQLParserEngineAdvice";
-    
-    private static final String JDBC_EXECUTOR_CALLBACK_ENGINE_ENHANCE_CLASS = "org.apache.shardingsphere.infra.executor.sql.execute.engine.driver.jdbc.JDBCExecutorCallback";
-    
-    private static final String JDBC_EXECUTOR_METHOD_NAME = "execute";
-    
-    private static final String JDBC_EXECUTOR_UNIT_ENGINE_ENHANCE_CLASS = "org.apache.shardingsphere.infra.executor.sql.execute.engine.driver.jdbc.JDBCExecutionUnit";
-    
-    private static final String JDBC_EXECUTOR_CALLBACK_ADVICE_CLASS = "org.apache.shardingsphere.agent.plugin.tracing.jaeger.advice.JDBCExecutorCallbackAdvice";
+    private final OpenTracingPluginDefinitionService pluginDefinitionService = new OpenTracingPluginDefinitionService();
     
     @Override
     public Collection<PluginInterceptorPoint> define() {
-        intercept(COMMAND_EXECUTOR_TASK_ENHANCE_CLASS)
-                .aroundInstanceMethod(ElementMatchers.named(COMMAND_EXECUTOR_METHOD_NAME))
-                .implement(COMMAND_EXECUTOR_TASK_ADVICE_CLASS)
-                .build();
-        intercept(SQL_PARSER_ENGINE_ENHANCE_CLASS)
-                .aroundInstanceMethod(ElementMatchers.named(SQL_PARSER_ENGINE_METHOD_NAME))
-                .implement(SQL_PARSER_ENGINE_ADVICE_CLASS)
-                .build();
-        intercept(JDBC_EXECUTOR_CALLBACK_ENGINE_ENHANCE_CLASS)
-                .aroundInstanceMethod(
-                        ElementMatchers.named(JDBC_EXECUTOR_METHOD_NAME)
-                                .and(ElementMatchers.takesArgument(0, ElementMatchers.named(JDBC_EXECUTOR_UNIT_ENGINE_ENHANCE_CLASS)))
-                )
-                .implement(JDBC_EXECUTOR_CALLBACK_ADVICE_CLASS)
-                .build();
-        return install();
+        return pluginDefinitionService.define();
     }
     
     @Override
