+++ shardingsphere-agent/shardingsphere-agent-plugins/shardingsphere-agent-plugin-tracing/shardingsphere-agent-tracing-opentracing/src/test/java/org/apache/shardingsphere/agent/plugin/tracing/opentracing/advice/SQLParserEngineAdviceTest.java
@@ -18,53 +18,32 @@
 package org.apache.shardingsphere.agent.plugin.tracing.opentracing.advice;
 
 import io.opentracing.mock.MockSpan;
-import io.opentracing.mock.MockTracer;
-import io.opentracing.util.GlobalTracer;
+import org.apache.shardingsphere.agent.api.result.MethodInvocationResult;
+import org.apache.shardingsphere.agent.plugin.tracing.advice.AbstractSQLParserEngineAdviceTest;
+import org.apache.shardingsphere.agent.plugin.tracing.opentracing.constant.OpenTracingConstants;
+import org.apache.shardingsphere.agent.plugin.tracing.rule.JaegerCollector;
+import org.junit.ClassRule;
+import org.junit.Test;
+
 import java.io.IOException;
-import java.lang.reflect.Method;
 import java.util.List;
 import java.util.Map;
-import org.apache.shardingsphere.agent.api.result.MethodInvocationResult;
-import org.apache.shardingsphere.agent.plugin.tracing.opentracing.constant.ErrorLogTagKeys;
-import org.apache.shardingsphere.infra.parser.ShardingSphereSQLParserEngine;
-
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.mockito.internal.util.reflection.FieldReader;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
-public final class SQLParserEngineAdviceTest {
+public final class SQLParserEngineAdviceTest extends AbstractSQLParserEngineAdviceTest {
+    
+    @ClassRule
+    public static JaegerCollector collector = new JaegerCollector();
     
     private static final SQLParserEngineAdvice ADVICE = new SQLParserEngineAdvice();
     
-    private static MockTracer tracer;
-    
-    private static Method parserMethod;
-    
-    @BeforeClass
-    public static void setup() throws NoSuchMethodException, NoSuchFieldException {
-        if (!GlobalTracer.isRegistered()) {
-            GlobalTracer.register(new MockTracer());
-        }
-        FieldReader fieldReader = new FieldReader(GlobalTracer.get(), GlobalTracer.class.getDeclaredField("tracer"));
-        tracer = (MockTracer) fieldReader.read();
-        parserMethod = ShardingSphereSQLParserEngine.class.getDeclaredMethod("parse", String.class, boolean.class);
-    }
-    
-    @Before
-    public void reset() {
-        tracer.reset();
-    }
-    
     @Test
     public void assertMethod() {
-        MockAdviceTargetObject targetObject = new MockAdviceTargetObject();
-        ADVICE.beforeMethod(targetObject, parserMethod, new Object[]{"select 1"}, new MethodInvocationResult());
-        ADVICE.afterMethod(targetObject, parserMethod, new Object[]{}, new MethodInvocationResult());
-        List<MockSpan> spans = tracer.finishedSpans();
+        ADVICE.beforeMethod(getTargetObject(), null, new Object[]{"select 1"}, new MethodInvocationResult());
+        ADVICE.afterMethod(getTargetObject(), null, new Object[]{}, new MethodInvocationResult());
+        List<MockSpan> spans = collector.finishedSpans();
         assertThat(spans.size(), is(1));
         assertThat(spans.get(0).logEntries().size(), is(0));
         assertThat(spans.get(0).operationName(), is("/ShardingSphere/parseSQL/"));
@@ -72,19 +51,18 @@ public final class SQLParserEngineAdviceTest {
     
     @Test
     public void assertExceptionHandle() {
-        MockAdviceTargetObject targetObject = new MockAdviceTargetObject();
-        ADVICE.beforeMethod(targetObject, parserMethod, new Object[]{"select 1"}, new MethodInvocationResult());
-        ADVICE.onThrowing(targetObject, parserMethod, new Object[]{}, new IOException());
-        ADVICE.afterMethod(targetObject, parserMethod, new Object[]{}, new MethodInvocationResult());
-        List<MockSpan> spans = tracer.finishedSpans();
+        ADVICE.beforeMethod(getTargetObject(), null, new Object[]{"select 1"}, new MethodInvocationResult());
+        ADVICE.onThrowing(getTargetObject(), null, new Object[]{}, new IOException());
+        ADVICE.afterMethod(getTargetObject(), null, new Object[]{}, new MethodInvocationResult());
+        List<MockSpan> spans = collector.finishedSpans();
         assertThat(spans.size(), is(1));
         MockSpan span = spans.get(0);
         assertThat(span.tags().get("error"), is(true));
         List<MockSpan.LogEntry> entries = span.logEntries();
         assertThat(entries.size(), is(1));
         Map<String, ?> fields = entries.get(0).fields();
-        assertThat(fields.get(ErrorLogTagKeys.EVENT), is("error"));
-        assertThat(fields.get(ErrorLogTagKeys.ERROR_KIND), is("java.io.IOException"));
+        assertThat(fields.get(OpenTracingConstants.ErrorLogTagKeys.EVENT), is("error"));
+        assertThat(fields.get(OpenTracingConstants.ErrorLogTagKeys.ERROR_KIND), is("java.io.IOException"));
         assertThat(span.operationName(), is("/ShardingSphere/parseSQL/"));
     }
 }
